Add NFS-mouting example.  Fix a couple of warnings, and reinstate -Wall.

+++ klibc/inet/inet_ntop.c
@@ -10,7 +10,7 @@
 
 const char *inet_ntop(int af, const void *cp, char *buf, size_t len)
 {
-  int xlen;
+  size_t xlen;
 
   switch ( af ) {
   case AF_INET:
